requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
